[PPT-14.1] Relocate ProtocolPreferences out of legacy protocol_grid (#419)
- New pluggable_protocol_tree/services/preferences.py exporting
  ProtocolPreferences, StepTime, ProtocolPreferencesPane and
  protocol_grid_tab — field names, preferences_path
  ("microdrop.protocol") and category id/ordering unchanged so
  persisted user settings keep loading.
- The three settling defaults move to pluggable_protocol_tree/consts.py.
- protocol_grid/preferences.py is now a thin re-export shim (until
  PPT-9 deletes the legacy plugin).
- PluggableProtocolTreePlugin contributes the pane + category; the
  legacy plugin no longer does (no duplicate Settings tab).
- Non-legacy importers re-pointed: capture_column (TODO(PPT-9) marker
  resolved), test_capture_column, test_persistence,
  dropbot_preferences_ui/preferences.
- Dropped the dead _level_default method (no matching trait existed).
